% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lesson_data.R
\name{get_lesson_data}
\alias{get_lesson_data}
\title{Get Lesson Data}
\usage{
get_lesson_data()
}
\value{
A data frame with lesson information
}
\description{
Returns the data frame containing information about all available lessons
}
\examples{
get_lesson_data()
}
